refactor(cache): split weather data serialization
Move persisted WeatherData serialization helpers out of cache.py while preserving private compatibility re-exports used by existing tests and callers.

Tested: uv run pytest -q -n 0 --tb=short tests/test_cache.py tests/test_cache_weather_alert_affected_zones.py tests/test_forecast_product_service.py tests/test_weather_client.py; uv run ruff check src tests scripts; uv run pyright

Confidence: high
